repository/backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
sqlalchemy[asyncio]>=2.0
aiomysql>=0.2
alembic>=1.13
passlib[bcrypt]>=1.7
python-jose[cryptography]>=3.3
pydantic-settings>=2.0
python-multipart>=0.0.9
httpx>=0.27
